fix(store): a bucket's range is a bound, and the hot ring is sorted (#1061, #1062)
min/max were f32 written with an `as` cast, which rounds to nearest rather
than outward — so the "range the bucket covered" was not a range. 2^24 + 1
stored as 2^24; at rx_bytes scale an f32 ulp is ~65 KB; and because `last`
stayed exact a bucket could report a max below its own last. agg=max is what
the range API sells on this number. Both are f64 now, SCHEMA_VERSION 4 -> 5.

The bump exposed the half that would have hurt: only the GUI moved a
wrong-schema file aside. The historian called open_with_cache directly and
treated any error as "run memory-only", so a schema bump would have cost every
historian its disk tiers permanently — silently, on every restart. Both
callers now go through PersistentStore::open_or_move_aside.

RingBuffer::push never compared timestamps while every reader promises
oldest-first. The historian ingests with recovery on, so a retransmitted
sample lands after samples newer than itself, counter_rate answers None on the
inverted pair, and the chart goes blank exactly when recovery had repaired the
gap it was drawing over. push now inserts at its place and returns what it
did; the historian counts reordered and too_old, the first reaching
@rpc/historian/stats as reordered_total. Neither had a counter before.

Tests first, red then green: a_bucket_bounds_the_value_it_holds,
a_stored_bucket_bounds_the_value_it_holds,
the_hot_ring_holds_samples_in_timestamp_order,
a_sample_older_than_a_full_ring_is_dropped_and_says_so,
a_v4_file_is_refused_by_schema_and_then_moved_aside,
a_late_sample_is_recorded_in_order_and_counted.

Part of #1052.

+++ zensight-historian/src/query/range.rs
@@ -235,8 +235,8 @@ pub fn reduce(points: &[(i64, Bucket)], step_s: i64, agg: Aggregate) -> Vec<(i64
     for (ts, b) in points {
         let step_ts = ts.div_euclid(step_ms) * step_ms;
         let v = match agg {
-            Aggregate::Min => b.min as f64,
-            Aggregate::Max => b.max as f64,
+            Aggregate::Min => b.min,
+            Aggregate::Max => b.max,
             _ => b.last,
         };
         match &mut current {
@@ -432,7 +432,7 @@ mod tests {
         RpcRequest::new(Vec::new(), params)
     }
 
-    fn b(last: f64, min: f32, max: f32) -> Bucket {
+    fn b(last: f64, min: f64, max: f64) -> Bucket {
         Bucket { last, min, max }
     }
